move states out of scanner
