+++ httpobs/__init__.py
@@ -1,2 +1,12 @@
 SOURCE_URL = 'https://github.com/mozilla/http-observatory'
 VERSION = '0.9.3'
+
+# The various statuses
+STATE_ABORTED = 'ABORTED'
+STATE_FAILED = 'FAILED'
+STATE_FINISHED = 'FINISHED'
+STATE_PENDING = 'PENDING'
+STATE_STARTING = 'STARTING'
+STATE_RUNNING = 'RUNNING'
+
+STATES = (STATE_ABORTED, STATE_FAILED, STATE_FINISHED, STATE_PENDING, STATE_RUNNING, STATE_STARTING)
