+++ httpobs/scanner/__init__.py
@@ -1,24 +1,2 @@
 # Current algorithm version
 ALGORITHM_VERSION = 2
-
-# The various statuses
-STATE_ABORTED = 'ABORTED'
-STATE_FAILED = 'FAILED'
-STATE_FINISHED = 'FINISHED'
-STATE_PENDING = 'PENDING'
-STATE_STARTING = 'STARTING'
-STATE_RUNNING = 'RUNNING'
-
-STATES = (STATE_ABORTED, STATE_FAILED, STATE_FINISHED, STATE_PENDING, STATE_RUNNING, STATE_STARTING)
-
-__all__ = [
-    'ALGORITHM_VERSION',
-    'NUM_TESTS',
-    'STATES',
-    'STATE_ABORTED',
-    'STATE_FAILED',
-    'STATE_FINISHED',
-    'STATE_PENDING',
-    'STATE_RUNNING',
-    'STATE_STARTING',
-]
